[Feat] Now all messages have [?] icon as interface to other programs [Fix]  Controlling files not found or directories

+++ src/com/jaime/ultransfer/network/TCPsocket.java
@@ -129,7 +129,7 @@ public class TCPsocket {
 
     }
 
-    public int getByte() {
+    public byte getByte() {
 
         byte nt;
